Check.php commented out
There are a few stray comments that I did not get rid of. Please check
if they are necessary and either remove or incorporate them.

I also forgot to mention in the last commit: Evelyn, I did not comment
on the JavaScript section so you'll have to put some in.